requirements.txt
Django==4.2.1
Pillow==9.5.0
django-debug-toolbar==3.6.0
django-decouple==2.1
django-mptt==0.14.0
celery==5.2.7